+++ cli-tests/Cargo.toml
@@ -7,7 +7,9 @@ edition = "2021"
 api = { path = "../api" }
 assert_cmd = "2.0.16"
 assert_matches = "1.5.0"
+bundle = { path = "../bundle" }
 chrono = "0.4.33"
+codeowners = { path = "../codeowners" }
 context = { path = "../context" }
 escargot = "0.5.12"
 junit-mock = { path = "../junit-mock" }
