+++ cli-tests/src/upload.rs
@@ -9,14 +9,12 @@ use api::{
 };
 use assert_cmd::Command;
 use assert_matches::assert_matches;
+use bundle::{BundleMeta, FileSetType};
+use codeowners::CodeOwners;
 use context::repo::RepoUrlParts as Repo;
 use predicates::prelude::*;
 use tempfile::tempdir;
 use test_utils::mock_server::{MockServerBuilder, RequestPayload};
-use trunk_analytics_cli::{
-    codeowners::CodeOwners,
-    types::{BundleMeta, FileSetType},
-};
 
 // NOTE: must be multi threaded to start a mock server
 #[tokio::test(flavor = "multi_thread")]
