+++ cli/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "trunk-analytics-cli"
 edition = "2021"
-version = "0.0.0"
+version = "0.1.0"
 
 [[bin]]
 name = "trunk-analytics-cli"
@@ -14,6 +14,8 @@ path = "src/lib.rs"
 [dependencies]
 anyhow = "1.0.44"
 api = { path = "../api" }
+bundle = { path = "../bundle" }
+constants = { path = "../constants" }
 chrono = { version = "0.4.33", default-features = false, features = ["clock"] }
 clap = { version = "4.4.18", features = ["derive", "env"] }
 context = { path = "../context" }
@@ -28,22 +30,16 @@ tokio = { version = "*", default-features = false, features = [
 tempfile = "3.2.0"
 tokio-retry = { version = "0.3", default-features = false }
 glob = "0.3.0"
-regex = { version = "1.10.3", default-features = false, features = ["std"] }
 reqwest = { version = "0.12.5", default-features = false, features = [
   "rustls-tls-native-roots",
   "stream",
   "json",
 ] }
-serde = { version = "1.0.130", default-features = false, features = ["derive"] }
-serde_json = "1.0.68"
-zstd = { version = "0.13.0", default-features = false }
-tar = { version = "0.4.30", default-features = false }
 codeowners = { path = "../codeowners" }
 xcresult = { path = "../xcresult" }
 sentry = { version = "0.34.0", features = ["debug-images"] }
 openssl = { version = "0.10.66", features = ["vendored"] }
 openssl-src = "=300.3.1+3.3.1"
-uuid = { version = "1.10.0", features = ["v5"] }
 quick-junit = "0.5.0"
 colored = "2.1.0"
 console = "0.15.8"
