+++ cli/src/lib.rs
@@ -1,10 +1,5 @@
 pub mod api_client;
-pub mod bundler;
-pub mod codeowners;
-pub mod constants;
 pub mod runner;
 pub mod scanner;
-pub mod types;
 pub mod upload;
-pub mod utils;
 pub mod validate;
