+++ codeowners/Cargo.toml
@@ -15,5 +15,9 @@ lazy_static = "1.4"
 anyhow = "1.0.86"
 thiserror = "1.0.63"
 const_format = "0.2.33"
+constants = { path = "../constants" }
 fancy-regex = "0.13.0"
 indexmap = "2.5.0"
+log = "0.4.14"
+serde = { version = "1.0.130", default-features = false, features = ["derive"] }
+serde_json = "1.0.68"
