+++ xcresult/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "xcresult"
 edition = "2021"
-version = "0.0.0"
+version = "0.1.0"
 
 [lib]
 name = "xcresult"
